Sletning af en sag implementeret. Fejlhåndtering og genindlæsning af den tidligere UserControl er tilføjet.

+++ Bol-IT/Bol-IT/UserControls/Sag_Edit.cs
@@ -440,7 +440,15 @@ namespace Bol_IT
             }
         }
 
+        //Tobias
         private void btnCancel_Click(object sender, EventArgs e)
+        {
+            ReloadPrievous();
+        }
+
+        //Caspar
+        //Genindlæser den forrige UserControl, og beder den om at genindlæse den releveante data.
+        private static void ReloadPrievous()
         {
             //Load Sag_ViewAll User control når tryk på knap
             if (!Form1.Instance.PnlContainer.Controls.ContainsKey("Sag_ViewAll"))
@@ -449,6 +457,7 @@ namespace Bol_IT
                 Form1.Instance.PnlContainer.Controls.Add(Sag_ViewAll.Instance);
             }
             Form1.Instance.PnlContainer.Controls["Sag_ViewAll"].BringToFront();
+            Sag_ViewAll.Instance.StartDataLoad();
         }
 
         //Christoffer
@@ -632,12 +641,25 @@ namespace Bol_IT
                 lvHouseFiles.Items.Remove(item);
             }
         }
+
+        //Caspar
+        //Sletter en ejendom fra databasen ved tryk på Slet knappen.
         private void btnDeleteCase_Click(object sender, EventArgs e)
         {
             try
             {
-                int.TryParse(rtbCaseNr.Text, out int caseNr);
-                DataAccessLayerFacade.DeleteProperty(caseNr);
+                if (MessageBox.Show("Sagen bliver slettet for evigt. Ønsker du at fortsætte?", "Slet sag.", MessageBoxButtons.YesNo, MessageBoxIcon.Information) == DialogResult.No)
+                {
+                    return;
+                }
+                else
+                {
+                    int.TryParse(rtbCaseNr.Text, out int caseNr);
+                    DataAccessLayerFacade.DeleteProperty(caseNr);
+                    ReloadPrievous();
+
+                }
+
             }
             catch (Exception exception)
             {
